+++ solvers/aggs/AggSolver.h
@@ -43,7 +43,7 @@ public:
 	virtual ~AggSolver();
 
 	/////////////////////SOLVER NECESSARY
-	void 	backtrack 		(const Lit& l);
+	void 		backtrack 		(const Lit& l);
 
 	/*
 	 * Returns the explanation for the deduction of p from an aggregate expression.
@@ -51,23 +51,23 @@ public:
 	 * @post the first element in the reason clause will be the literal itself (invariant by minisat!)
 	 * @post the clause is not saved, so HAS to be deleted after use
 	 */
-	Clause* getExplanation	(const Lit& p);
-	void 	notifyVarAdded	(uint nvars); 		//correctly initialize AMNSolver datastructures when vars are added
-	Clause* propagate	(const Lit& p);
+	Clause* 	getExplanation	(const Lit& p);
+	void 		notifyVarAdded	(uint64_t nvars); 		//correctly initialize AMNSolver datastructures when vars are added
+	Clause* 	propagate	(const Lit& p);
 	/////////////////////ENDSOLVER NECESSARY
 
 	/////////////////////IDSOLVER NECESSARY
-	void propagateJustifications(Lit l, vec<vec<Lit> >& jstf, vec<Lit>& v, vec<int> &nb_body_lits_to_justify);
+	void 		propagateJustifications(Lit l, vec<vec<Lit> >& jstf, vec<Lit>& v, vec<int> &nb_body_lits_to_justify);
 
-	bool findJustificationAggr(Var head, vec<Lit>& jstf);
+	bool 		findJustificationAggr(Var head, vec<Lit>& jstf);
 
-	bool directlyJustifiable(Var v, vec<Lit>& jstf, vec<Var>& nonjstf, vec<Var>& currentjust);
-	void createLoopFormula(Var v, const std::set<Var>& ufs, vec<Lit>& loopf, vec<int>& seen);
+	bool 		directlyJustifiable(Var v, vec<Lit>& jstf, vec<Var>& nonjstf, vec<Var>& currentjust);
+	void 		createLoopFormula(Var v, const std::set<Var>& ufs, vec<Lit>& loopf, vec<int>& seen);
 
 	/**
 	 * Adds the heads of the aggregates in which x occurs as a body literal to heads
 	 */
-	void getHeadsOfAggrInWhichOccurs(Var x, vec<Var>& heads);
+	void 		getHeadsOfAggrInWhichOccurs(Var x, vec<Var>& heads);
 
 	/**
 	 * Returns the set literals of the aggregate with the given head x.
@@ -88,7 +88,7 @@ public:
 	 *
 	 * @remark: please ensure that all id numbers are used without gaps in the numbering.
 	 */
-	bool    addSet       (int id, vec<Lit>& l, vector<Weight>& w);
+	bool    	addSet       (int id, vec<Lit>& l, vector<Weight>& w);
 
 	/**
 	 * Adds an aggregate of the given type with number defn for set set_id.
@@ -101,13 +101,13 @@ public:
 	 *
 	 * @pre: no weights==0 when using a product aggregate
 	 */
-	bool    addAggrExpr  (int defn, int set_id, Weight bound, bool lower, AggrType type, bool defined);
+	bool    	addAggrExpr  (int defn, int set_id, Weight bound, bool lower, AggrType type, bool defined);
 
 	/**
 	 * Checks presence of aggregates and initializes all counters.
 	 * @Return: true if there are aggregates present
 	 */
-	bool    finishECNF_DataStructures (); //throws UNSAT
+	bool    	finishECNF_DataStructures (); //throws UNSAT
 
 	pPCSolver	getSolver				()	const		{ return solver; }
 	//void		resetIDSolver			()				{ idsolver.reset(); }
@@ -117,25 +117,25 @@ public:
 	/////////////////////END INITIALIZATION
 
 	//are used by agg.c, but preferably should be move into protected again
-	Clause* notifySATsolverOfPropagation(const Lit& p, Aggrs::AggrReason* cr);	// Like "enqueue", but for aggregate propagations.
+	Clause* 	notifySATsolverOfPropagation(const Lit& p, Aggrs::AggrReason* cr);	// Like "enqueue", but for aggregate propagations.
 
 	//Optimisation support
-	bool 	addMnmzSum		(Var headv, int setid, bool lower);
-    bool 	invalidateSum	(vec<Lit>& invalidation, Var head);
-    void 	propagateMnmz	(Var head);
+	bool 		addMnmzSum		(Var headv, int setid, bool lower);
+    bool 		invalidateSum	(vec<Lit>& invalidation, Var head);
+    void 		propagateMnmz	(Var head);
 
 protected:
 	/**
 	 * Returns the aggregate in which the given variable is the head.
 	 */
-    pAgg	getAggWithHeadOccurence	(Var v) const;
+    pAgg		getAggWithHeadOccurence	(Var v) const;
 
 	vector<pSet>	aggrminsets;
 	vector<pSet>	aggrmaxsets;
 	vector<pSet>	aggrsumsets;
 	vector<pSet>	aggrprodsets;
 
-	vector<AggrReason*>	aggr_reason;	// For each atom, like 'reason'.
+	vector<AggrReason*>			aggr_reason;	// For each atom, like 'reason'.
 	vector<vector<AggrWatch> >	aggr_watches;	// Aggr_watches[v] is a list of sets in which VAR v occurs (each AggrWatch says: which set, what type of occurrence).
 
 	//index on VAR (heads are always positive
@@ -153,17 +153,17 @@ protected:
 	 *
 	 * @PRE: backtracking is in anti-chronologous order and all literals are visited!
 	 */
-	void 	doBacktrack(const Lit& l);
+	void 		doBacktrack(const Lit& l);
 
 	bool		init;	//indicates whether still in initialization mode
 
 	/**
 	 * Goes through all watches and propagates the fact that p was set true.
 	 */
-	Clause* Aggr_propagate		(const Lit& p);
+	Clause* 	Aggr_propagate		(const Lit& p);
 
-	bool 	maxAggAsSAT(bool defined, bool lower, Weight bound, const Lit& head, const AggrSet& set);
-	bool	finishSets(vector<pSet>& sets); //throws UNSAT
+	bool 		maxAggAsSAT(bool defined, bool lower, Weight bound, const Lit& head, const AggrSet& set);
+	bool		finishSets(vector<pSet>& sets); //throws UNSAT
 };
 
 //=======================
